Add files via upload
re-commiting because I forgot to comment on the first.

Added some hover feature to buttons when cursor is moved over them as well as the cursor changing to a pointer when over top of the 
buttons. Buttons also have links in them so when they are clicked they do something. Currently they open a new tab and go to google. Going to change this soon.